src/requirements.txt
locustio==0.14.6
pyjwt>=1.7.1
cryptography>=2.9.2